+++ Modules/Microsoft365DSC/DSCResources/MSFT_AADRoleDefinition/MSFT_AADRoleDefinition.psm1
@@ -115,6 +115,7 @@ function Get-TargetResource
                 Ensure                = "Present"
                 Credential            = $Credential
                 ApplicationId         = $ApplicationId
+                ApplicationSecret     = $ApplicationSecret
                 TenantId              = $TenantId
                 CertificateThumbprint = $CertificateThumbprint
             }
@@ -399,6 +400,7 @@ function Export-TargetResource
             $Params = @{
                 Credential            = $Credential
                 ApplicationId         = $ApplicationId
+                ApplicationSecret     = $ApplicationSecret
                 TenantId              = $TenantId
                 CertificateThumbprint = $CertificateThumbprint
                 DisplayName           = $AADRoleDefinition.DisplayName