+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 # Change log for Microsoft365DSC
 
+# 1.21.1124.1
+
+* IntuneAppProtectionPolicyiOS
+  * Fixes to the Invoke-MgGraphRequest cmdlets parameters;
+* MISC
+  * Removed the Quiet switch from the Assert Blueprint cmdlet;
+    ISSUE #1563
+* DEPENDENCIES
+  * Updated Microsoft.Graph.Applications to 1.9.0;
+  * Updated Microsoft.Graph.Authentication to 1.9.0;
+  * Updated Microsoft.Graph.DeviceManagement to 1.9.0;
+  * Updated Microsoft.Graph.DeviceManagement.Administration to 1.9.0;
+  * Updated Microsoft.Graph.DeviceManagement.Enrolment to 1.9.0;
+  * Updated Microsoft.Graph.Devices.CorporateManagement to 1.9.0;
+  * Updated Microsoft.Graph.Groups to 1.9.0;
+  * Updated Microsoft.Graph.Identity.DirectoryManagement to 1.9.0;
+  * Updated Microsoft.Graph.Identity.SignIns to 1.9.0;
+  * Updated Microsoft.Graph.Planner to 1.9.0;
+  * Updated Microsoft.Graph.Teams to 1.9.0;
+  * Updated Microsoft.Graph.Users to 1.9.0;
+  * Updated MSCloudLoginAssistant to 1.0.79;
+  * Updated ReverseDSC to 2.0.0.8;
+
 # 1.21.1117.2
 
 * MISC